Refactor: Modularize visualize_fields
